Move notifications from team
