+++ src/frontend/static/css/podcastmanagement/podcastmanagement.css
@@ -1040,7 +1040,7 @@ body {
 }
 
 /* Notification */
-.notification {
+/* .notification {
   position: fixed;
   top: 16px;
   right: 16px;
@@ -1114,7 +1114,7 @@ body {
 
 .notification-close:hover {
   color: var(--highlight-color);
-}
+} */
 
 /* Popup Styles */
 .popup {
