+++ src/frontend/static/js/account/account.js
@@ -5,6 +5,7 @@ import {
   deleteUserAccount,
   subscribeUser,
 } from "/static/requests/accountRequests.js"
+import { showNotification } from "../components/notifications.js"
 
 document.addEventListener("DOMContentLoaded", () => {
     // Hides the edit buttons when in non-edit mode
@@ -36,12 +37,12 @@ document.addEventListener("DOMContentLoaded", () => {
 
         // If name is missing, show a message
         if (!data.full_name) {
-          showNotification("Please complete your profile by adding your full name", "warning")
+          showNotification("Info", "Please complete your profile by adding your full name", "info")
         }
 
         // If phone is missing, show a message
         if (!data.phone) {
-          showNotification("Please add your phone number to complete your profile", "warning")
+          showNotification("Info", "Please add your phone number to complete your profile", "info")
         }
       } else {
         console.error("Error fetching profile data")
@@ -534,7 +535,7 @@ function triggerFileUpload() {
   document.body.removeChild(fileInput)
 }
 
-function showNotification(message, type) {
+function dshowNotification(message, type) {
   const notificationContainer = document.getElementById("header-notifications");
   if (!notificationContainer) return;
