+++ src/frontend/templates/components/notification.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Components</title>
+    <!-- Link to the CSS file -->
+    <link rel="stylesheet" href="../../static/css/components/notifications.css">
+</head>
+<body>
+    <!-- Notification Component -->
+    <div id="notification-template" class="notification">
+        <div class="notification-icon"></div>
+        <div class="notification-content">
+            <div class="notification-title"></div>
+            <div class="notification-message"></div>
+        </div>
+        <div class="notification-close"></div>
+    </div>
+    
+    <!-- Confirmation Popup Component -->
+    <div id="confirmation-popup-template" class="popup confirmation-popup">
+        <div class="form-box">
+            <h2 class="form-title"></h2>
+            <p class="form-message"></p>
+            <div class="form-actions">
+                <button class="cancel-btn" id="cancelPopupBtn">Cancel</button>
+                <button class="save-btn" id="confirmPopupBtn">Confirm</button>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
