+++ src/requirements.txt
@@ -1,34 +1,166 @@
+absl-py
+aiohappyeyeballs
+altair
+annotated-types
+antlr4-python3-runtime
+anyio
 APScheduler
-azure_storage
+attrs
+audioread
+azure-cognitiveservices-speech
+bcrypt
+beautifulsoup4
+blinker
+cachetools
+certifi
+cffi
+charset-normalizer
+click
+cmudict
+colorama
+contourpy
+cycler
+decorator
+distro
 dnspython
+docopt
+durationpy
+einops
 elevenlabs
 Faker
 feedparser
+filelock
 Flask
-flask_cors
-flask_mail
-google_auth_oauthlib
+Flask-Cors
+Flask-Mail
+flatbuffers
+fonttools
+frozenlist
+fsspec
+gitdb
+GitPython
+google
+google-auth
+google-auth-oauthlib
+greenlet
+gunicorn
+h11
+h3
+httpcore
+httplib2
+httpx
+huggingface-hub
+idna
+importlib_metadata
+importlib_resources
+itsdangerous
+Jinja2
+jiter
+joblib
+jsonify
+jsonschema
+jsonschema-specifications
+kiwisolver
+kubernetes
+llvmlite
+Mako
+MarkupSafe
 marshmallow
 matplotlib
-mediapipe
+more-itertools
+mpmath
+multidict
+narwhals
+networkx
+nltk
+node
 noisereduce
+npm
+numba
 numpy
+oauth2client
+oauthlib
+odict
 openai
-openai_whisper
-opencv_contrib_python
-Pillow
+openai-whisper
+opencv-contrib-python
+opt_einsum
+optional-django
+packaging
+pandas
+php
+pillow
+plumber
+propcache
+protobuf
+pyarrow
+pyasn1
+pyasn1_modules
+pycparser
+pydantic
+pydantic_core
+pydeck
 pydub
+pylogrus
 pymongo
+pyparsing
+pyphen
+python-bootstrap
+python-dateutil
 python-dotenv
 pytz
+PyYAML
+referencing
+regex
+requests
+requests-oauthlib
+rpds-py
+rsa
+ruamel.yaml
+ruamel.yaml.clib
+safetensors
+scipy
+setuptools
+sgmllib3k
+shortuuid
+six
+smmap
+sniffio
 soundfile
+soupsieve
+sqlparse
 streamlit
+sympy
+tenacity
 textblob
 textstat
+tiktoken
 timezonefinder
+tokenizers
+toml
 torch
 torchaudio
 torchvision
+tornado
+tqdm
 transformers
+typing-inspection
+typing_extensions
+tzdata
+tzlocal
+urllib3
+watchdog
+websocket-client
+websockets
 Werkzeug
 whisper
+yarl
+zipp
+zope.component
+zope.deferredimport
+zope.deprecation
+zope.event
+zope.hookable
+zope.interface
+zope.lifecycleevent
+zope.proxy
